metadata.md - Reorganize for clarity.
- Reorganize to make it easier to follow top to bottom instead of having
  to skip around to complete each section.
- Standardize things to always match the actual keys used in source.
- Make the 'Notes' a little less scary-looking.

As a first-time reader, the section ordering seemed a little arbitrary.
This is an attempt to put the sections in an order which flows better
from the implementor perspective.

Some section have internal reorganization, as well.  In particular,
those with enumerated choices have the list presented earlier in the
discussion. The first sentence is always a description of the intent of
the bit of metadata. The last bit is always the example snippet.