location_api: reorganize for better flow.
- Explain why they might need it.
- Describe how they come to have it.
- Tell them what's available.
- Show them how to use it.
- Note development differences.

The "Is it snowing?" example code no longer aligns with the live
code, so just repurpose it as the example of what they will get
in the `duckpan` environment.

Reorganize and tidy up the example data to make it match the order of
attributes in the displayed code and align for easier scanning.

Add a link to how to use in testing.