Fix links to anchors in DuckDuckHack's Handlebars helper for spices
The trailing slashes in the absolute URLs did not link to the correct
anchor, and confused the navigation menu on the left of the pages on
duckduckhack. Also, some references used upper-case characters, which
should be all lowercase. Hence, we convert them to proper relative
links to the lowercase anchors, which makes the links work, and no
longer messes with the navigation menu's relative links.

Fixes #217